final/requirements.txt
flask
gunicorn
requests
google-auth
google-api-python-client
google-cloud-secret-manager
